flask/requirements.txt
Flask
pymongo
Pillow
tensorflow
nltk
